refactor: adopt src/englab package layout to fix test imports
