+++ pyproject.toml
@@ -1,0 +1,17 @@
+[project]
+name = "engineering-lab"
+version = "0.0.1"
+description = "Personal engineering lab for experiments."
+
+[tool.pytest.ini_options]
+addopts = "-q"
+
+[tool.ruff]
+line-length = 100
+target-version = "py311"
+select = ["E", "F", "I", "UP", "B", "N", "PL", "RUF"]
+ignore = ["E203"]
+src = ["src", "tests"]
+
+[tool.ruff.format]
+quote-style = "double"
