+++ pyrightconfig.json
@@ -1,0 +1,6 @@
+{
+  "include": ["src", "tests"],
+  "pythonVersion": "3.11",
+  "typeCheckingMode": "standard",
+  "reportMissingImports": true
+}
