+++ tests/test_broken_math.py
@@ -1,4 +1,4 @@
-from src.broken_math import add
+from englab.broken_math import add
 
 def test_addition_should_fail():
     assert add(1, 2) == 99  # failing on purpose
